fix "Tests should include assertions" issues
