+++ cxx-sensors/src/test/java/org/sonar/cxx/sensors/utils/CxxReportSensor_getReports_Test.java
@@ -23,12 +23,7 @@ import org.sonar.cxx.sensors.utils.CxxReportSensor;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.apache.commons.io.FileUtils;
@@ -46,72 +41,6 @@ public class CxxReportSensor_getReports_Test {
   @Rule
   public TemporaryFolder base = new TemporaryFolder();
 
-  @Test
-  public void getReports_patternMatching() throws java.io.IOException, java.lang.InterruptedException {
-    List<String[]> examples = new LinkedList<>();
-
-    //                        "pattern",      "matches",   "matches not"
-    // relative
-    examples.add(new String[]{"A.ext",        "A.ext",     "dir/B.ext"});
-    examples.add(new String[]{"dir/A.ext",    "dir/A.ext", "A.ext,dir/B.ext"});
-    examples.add(new String[]{"dir/../A.ext", "A.ext",     "B.ext,dir/A.ext"});
-    examples.add(new String[]{"./A.ext",      "A.ext",     "B.ext"});
-    examples.add(new String[]{"./A.ext",      "A.ext",     "B.ext"});
-    // empty
-    examples.add(new String[]{"", "", ""});
-    // question mark glob
-    examples.add(new String[]{"A?.ext",       "AA.ext,AB.ext", "B.ext"});
-    // multi-char glob
-    examples.add(new String[]{"A*.ext",       "A.ext,AAA.ext", "B.ext"});
-    // multi-dir glob
-    examples.add(new String[]{"**/A.ext",     "A.ext,dir/A.ext,dir/subdir/A.ext", "B.ext,dir/B.ext"});
-    examples.add(new String[]{"dir/**/A.ext", "dir/A.ext,dir/subdir/A.ext",       "A.ext,dir/B.ext,dir/subdir/B.ext"});
-
-    String pattern, match, allpaths;
-    List<File> reports;
-    for (String[] example : examples) {
-      pattern = example[0];
-      match = example[1];
-      allpaths = String.join(",", Arrays.copyOfRange(example, 1, 3));
-      setupExample(allpaths);
-
-      settings.setProperty(REPORT_PATH_KEY, pattern);
-      reports = CxxReportSensor.getReports(settings.asConfig(), base.getRoot(), REPORT_PATH_KEY);
-      assertMatch(reports, match, example[0]);
-      deleteExample(base.getRoot());
-    }
-
-    // TODO: some special windows cases?
-  }
-
-  private void setupExample(String pathes) throws java.io.IOException {
-    String[] parsedPaths = pathes.split(",");
-    for (String path : parsedPaths) {
-      path = path.trim();
-      if (!path.isEmpty()) {
-        FileUtils.touch(new File(base.getRoot(), path));
-      }
-    }
-  }
-
-  private void deleteExample(File dir) throws java.io.IOException {
-    FileUtils.cleanDirectory(dir);
-  }
-
-  private void assertMatch(List<File> real, String expected, String pattern) {
-    String[] parsedPaths = expected.split(",");
-    List<File> expectedFiles = new LinkedList<>();
-    for (String path : parsedPaths) {
-      path = path.trim();
-      if (!path.isEmpty()) {
-        expectedFiles.add(new File(base.getRoot(), path));
-      }
-    }
-
-    Set<File> realSet = new TreeSet<>(real);
-    Set<File> expectedSet = new TreeSet<>(expectedFiles);
-    assertThat(realSet).describedAs("Failed for pattern: {}", pattern).isEqualTo(expectedSet);
-  }
 
   @Test
   public void testAbsoluteInsideBasedir() throws IOException {
